+++ README.md
@@ -3,7 +3,7 @@ Python wrapper around MetOffice Atmospheric Model Data REST API
 
 [![codecov](https://codecov.io/gh/openclimatefix/MetOfficeAMD/branch/main/graph/badge.svg?token=64JOBKZNCI)](https://codecov.io/gh/openclimatefix/MetOfficeAMD)
 
-MetOfficeAMD is a simple wrapper for the API provided by the British Met Office <https://metoffice.apiconnect.ibmcloud.com/metoffice/production/>_ known as Weather DataHub. It can be used to retrieve weather observations and forecasts. The aim is to forecus on the AMD - Atmosphere Model Data REST API <https://metoffice.apiconnect.ibmcloud.com/metoffice/production/product/17502/api/16908>
+MetOfficeAMD is a simple wrapper for the API provided by the British Met Office <https://metoffice.apiconnect.ibmcloud.com/metoffice/production/>_ known as Weather DataHub. It can be used to retrieve weather observations and forecasts. The aim is to focus on the AMD - Atmosphere Model Data REST API <https://metoffice.apiconnect.ibmcloud.com/metoffice/production/product/17502/api/16908>
 
 ## Access
 Your need to register and obtain a API key and secret. This should be placed in environment variables as `API_KEY` and `API_SECRET`
